+++ src/textual_textarea/text_editor.py
@@ -1,12 +1,11 @@
 from __future__ import annotations
 
 import re
-from collections import deque
 from contextlib import suppress
 from dataclasses import dataclass
 from math import ceil, floor
 from pathlib import Path
-from typing import TYPE_CHECKING, Any, Callable, Deque, Literal, Sequence
+from typing import TYPE_CHECKING, Any, Callable, Literal, Sequence
 
 import pyperclip
 from rich.console import RenderableType
@@ -42,7 +41,6 @@ BRACKETS = {
     "{": "}",
 }
 CLOSERS = {'"': '"', "'": "'", **BRACKETS}
-UNDO_SIZE = 25
 
 # these patterns need to match a reversed string!
 DOUBLE_QUOTED_EXPR = r'"([^"\\]*(\\.[^"\\]*|""[^"\\]*)*)"(b?r|f|b|rb|&?u|@)?'
@@ -86,8 +84,8 @@ class TextAreaPlus(TextArea, inherit_bindings=False):
         Binding("ctrl+right", "cursor_word_right", "cursor word right", show=False),
         Binding("home", "cursor_line_start", "cursor line start", show=False),
         Binding("end", "cursor_line_end", "cursor line end", show=False),
-        Binding("ctrl+home", "cursor_doc_start", "cursor line start", show=False),
-        Binding("ctrl+end", "cursor_doc_end", "cursor line end", show=False),
+        Binding("ctrl+home", "cursor_doc_start", "cursor doc start", show=False),
+        Binding("ctrl+end", "cursor_doc_end", "cursor doc end", show=False),
         Binding("pageup", "cursor_page_up", "cursor page up", show=False),
         Binding("pagedown", "cursor_page_down", "cursor page down", show=False),
         # scrolling
@@ -114,6 +112,18 @@ class TextAreaPlus(TextArea, inherit_bindings=False):
         ),
         Binding(
             "shift+end", "cursor_line_end(True)", "cursor line end select", show=False
+        ),
+        Binding(
+            "ctrl+shift+home",
+            "cursor_doc_start(True)",
+            "select to cursor doc start",
+            show=False,
+        ),
+        Binding(
+            "ctrl+shift+end",
+            "cursor_doc_end(True)",
+            "select to cursor doc end",
+            show=False,
         ),
         Binding("shift+up", "cursor_up(True)", "cursor up select", show=False),
         Binding("shift+down", "cursor_down(True)", "cursor down select", show=False),
@@ -198,8 +208,6 @@ class TextAreaPlus(TextArea, inherit_bindings=False):
         )
         self.cursor_blink = False if self.app.is_headless else True
         self.use_system_clipboard = use_system_clipboard
-        self.undo_stack: Deque[InputState] = deque(maxlen=UNDO_SIZE)
-        self.redo_stack: Deque[InputState] = deque(maxlen=UNDO_SIZE)
         self.double_click_location: Location | None = None
         self.double_click_timer: Timer | None = None
         self.consecutive_clicks: int = 0
@@ -207,24 +215,12 @@ class TextAreaPlus(TextArea, inherit_bindings=False):
         self.system_paste: Callable[[], str] | None = None
 
     def on_mount(self) -> None:
-        self.undo_timer = self.set_interval(
-            interval=0.3,
-            callback=self._create_undo_snapshot,
-            name="undo_timer",
-            pause=not self.has_focus,
-        )
         self._determine_clipboard()
-        self._create_undo_snapshot()
 
     def on_blur(self, event: events.Blur) -> None:
         event.prevent_default()
         self._pause_blink(visible=False)
         self.post_message(TextAreaHideCompletionList())
-        self._create_undo_snapshot()
-        self.undo_timer.pause()
-
-    def on_focus(self) -> None:
-        self.undo_timer.reset()
 
     def on_key(self, event: events.Key) -> None:
         # Naked shift or ctrl keys on Windows get sent as NUL chars; Textual
@@ -236,7 +232,6 @@ class TextAreaPlus(TextArea, inherit_bindings=False):
             event.prevent_default()
             return
 
-        self.undo_timer.reset()
         if event.key in (
             "apostrophe",
             "quotation_mark",
@@ -270,7 +265,6 @@ class TextAreaPlus(TextArea, inherit_bindings=False):
             self.post_message(TextAreaHideCompletionList())
 
     def on_mouse_down(self, event: events.MouseDown) -> None:
-        self.undo_timer.reset()
         self.post_message(TextAreaHideCompletionList())
         target = self.get_target_document_location(event)
         if (
@@ -298,7 +292,6 @@ class TextAreaPlus(TextArea, inherit_bindings=False):
                 self.action_select_all()
             self.consecutive_clicks += 1
         else:
-            self._create_undo_snapshot()
             self.double_click_location = target
             self.consecutive_clicks += 1
 
@@ -313,7 +306,6 @@ class TextAreaPlus(TextArea, inherit_bindings=False):
         event.prevent_default()
         event.stop()
         self.post_message(TextAreaHideCompletionList())
-        self._create_undo_snapshot()
         self.replace(event.text, *self.selection, maintain_selection_offset=False)
 
     @on(ClipboardReady)
@@ -345,25 +337,29 @@ class TextAreaPlus(TextArea, inherit_bindings=False):
 
     def action_cut(self) -> None:
         self.post_message(TextAreaHideCompletionList())
-        self._create_undo_snapshot()
         self._copy_selection()
         if not self.selected_text:
             self.action_delete_line()
         self.delete(*self.selection)
 
-    def action_cursor_doc_start(self) -> None:
+    def action_cursor_doc_start(self, select: bool = False) -> None:
         self.post_message(TextAreaHideCompletionList())
-        self.selection = Selection(start=(0, 0), end=(0, 0))
+        if select:
+            self.selection = Selection(start=self.selection.start, end=(0, 0))
+        else:
+            self.selection = Selection(start=(0, 0), end=(0, 0))
 
-    def action_cursor_doc_end(self) -> None:
+    def action_cursor_doc_end(self, select: bool = False) -> None:
         self.post_message(TextAreaHideCompletionList())
         lno = self.document.line_count - 1
         loc = (lno, len(self.document.get_line(lno)))
-        self.selection = Selection(start=loc, end=loc)
+        if select:
+            self.selection = Selection(start=self.selection.start, end=loc)
+        else:
+            self.selection = Selection(start=loc, end=loc)
 
     def action_delete_line(self) -> None:
         self.post_message(TextAreaHideCompletionList())
-        self._create_undo_snapshot()
         if self.selection.start != self.cursor_location:  # selection active
             self.delete(*self.selection, maintain_selection_offset=False)
         else:
@@ -447,36 +443,29 @@ class TextAreaPlus(TextArea, inherit_bindings=False):
                     )
             # add comment tokens to all lines
             else:
-                indent = min(
+                comment_indent = min(
                     [indent for indent, line in zip(indents, stripped_lines) if line]
                 )
+                insertion = f"{self.inline_comment_marker} "
                 for lno, stripped_line in enumerate(stripped_lines, start=first[0]):
                     if stripped_line:
-                        self.insert(
-                            f"{self.inline_comment_marker} ",
-                            location=(lno, indent),
-                            maintain_selection_offset=True,
-                        )
+                        # insert one character at a time, to create a single undo-able
+                        # batch of edits.
+                        # See https://github.com/Textualize/textual/issues/4428
+                        for i, char in enumerate(insertion):
+                            self.insert(
+                                char,
+                                location=(lno, comment_indent + i),
+                                maintain_selection_offset=True,
+                            )
 
     def action_undo(self) -> None:
         self.post_message(TextAreaHideCompletionList())
-        self._create_undo_snapshot()
-        if len(self.undo_stack) > 1:
-            # we just took a snapshot, so the current state is
-            # on the stack.
-            current_state = self.undo_stack.pop()
-            self.redo_stack.append(current_state)
-            prev_state = self.undo_stack[-1]
-            self.text = prev_state.text
-            self.selection = prev_state.selection
+        super().action_undo()
 
     def action_redo(self) -> None:
         self.post_message(TextAreaHideCompletionList())
-        if self.redo_stack:
-            state = self.redo_stack.pop()
-            self.undo_stack.append(state)
-            self.text = state.text
-            self.selection = state.selection
+        super().action_redo()
 
     def _clear_double_click(self) -> None:
         self.consecutive_clicks = 0
@@ -498,22 +487,6 @@ class TextAreaPlus(TextArea, inherit_bindings=False):
             except pyperclip.PyperclipException:
                 # no system clipboard; common in CI runners
                 self.post_message(TextAreaClipboardError(action="copy"))
-
-    def _create_undo_snapshot(self) -> None:
-        self.undo_timer.reset()
-        new_snapshot = InputState(
-            text=self.text,
-            selection=self.selection,
-        )
-        if self.undo_stack and self.undo_stack[-1] == new_snapshot:
-            return
-        elif self.undo_stack and self.undo_stack[-1].text == new_snapshot.text:
-            # overwrite the last checkpoint to just update the cursor
-            self.undo_stack[-1] = new_snapshot
-        else:
-            self.undo_stack.append(new_snapshot)
-            if self.redo_stack:
-                self.redo_stack = deque(maxlen=UNDO_SIZE)
 
     def _get_character_at_cursor(self) -> str:
         if self.cursor_at_end_of_line:
@@ -977,7 +950,7 @@ class TextEditor(Widget, can_focus=True, can_focus_children=False):
         """
         self.text_input.replace(
             text,
-            *list(sorted(self.text_input.selection)),
+            *self.text_input.selection,
             maintain_selection_offset=False,
         )